Refactor flutter app folder structure to group related screens and widgets
