+++ apps/flutter_app/lib/screens/auth/auth_wrapper.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import '../providers/auth_provider.dart';
-import 'auth/login_screen.dart';
-import 'home_screen.dart';
+import '../../providers/auth_provider.dart';
+import 'login_screen.dart';
+import '../home/home_screen.dart';
 
 class AuthWrapper extends StatelessWidget {
   const AuthWrapper({super.key});
