+++ apps/flutter_app/lib/screens/workspaces/workspace_list_screen.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import '../core/theme/valora_colors.dart';
-import '../core/theme/valora_typography.dart';
-import '../core/theme/valora_spacing.dart';
-import '../models/workspace.dart';
-import '../providers/workspace_provider.dart';
-import '../widgets/valora_widgets.dart';
-import '../widgets/workspaces/workspace_list_item.dart';
+import '../../core/theme/valora_colors.dart';
+import '../../core/theme/valora_typography.dart';
+import '../../core/theme/valora_spacing.dart';
+import '../../models/workspace.dart';
+import '../../providers/workspace_provider.dart';
+import '../../widgets/valora_widgets.dart';
+import '../../widgets/workspaces/workspace_list_item.dart';
 
 enum SortOption { name, createdDate, memberCount }
