+++ apps/flutter_app/lib/widgets/valora_widgets.dart
@@ -14,3 +14,6 @@ export 'common/valora_avatar.dart';
 export 'common/valora_search_field.dart';
 export 'common/valora_slider.dart';
 export 'common/valora_section_header.dart';
+export 'common/valora_glass_container.dart';
+export 'common/global_error_widget.dart';
+export 'common/valora_error_state.dart';
