+++ apps/flutter_app/test/screens/startup_screen_test.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:provider/provider.dart';
 import 'package:valora_app/providers/auth_provider.dart';
-import 'package:valora_app/screens/startup_screen.dart';
-import 'package:valora_app/widgets/valora_error_state.dart';
+import 'package:valora_app/screens/startup/startup_screen.dart';
+import 'package:valora_app/widgets/common/valora_error_state.dart';
 
 class DelayedAuthProvider extends ChangeNotifier implements AuthProvider {
   DelayedAuthProvider({required this.delay});
