+++ apps/flutter_app/test/widgets/workspaces/workspace_list_screen_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:provider/provider.dart';
 import 'package:valora_app/providers/workspace_provider.dart';
-import 'package:valora_app/screens/workspace_list_screen.dart';
+import 'package:valora_app/screens/workspaces/workspace_list_screen.dart';
 import 'package:valora_app/models/workspace.dart';
 import 'package:valora_app/models/saved_property.dart';
 import 'package:valora_app/models/activity_log.dart';
